not use half precision

+++ vigc/models/wav2vec_bert/bengali_t5_for_generation.py
@@ -165,7 +165,7 @@ class BengaliT5Transformer(BaseModel):
             self.load_checkpoint(url_or_filename=finetune_path)
             logging.info(f"Loaded finetuned model '{finetune_path}'.")
 
-    def maybe_autocast(self, dtype=torch.float16):
+    def maybe_autocast(self, dtype=torch.float32):
         # if on cpu, don't use autocast
         # if on gpu, use autocast with dtype if provided, otherwise use torch.float16
         enable_autocast = self.device != torch.device("cpu")
